feat : todos 경로 세그먼트에 error.tsx 파일 추가 및 서버 액션 에러 핸들링
- todos 경로 세그먼트에 error.tsx 파일을 추가하여 투두리스트 페이지
  렌더링 중 예상치 못한 에러 발생 시 에러 경계가 포착하여 error.tsx
  파일의 UI를 보여주도록 하였습니다.
- 서버 액션에 try ~ catch 구문을 추가하고 서버 액션에서 발생한 에러를
  반환하는 동작을 추가했습니다.

+++ src/actions/deleteTodo.ts
@@ -12,30 +12,44 @@ export const deleteTodo = async (
 ) => {
   const todoid = formData.get("todoid") as string;
 
-  const db = (await connectDB).db("next-todo-chart-cluster");
-  const todoDoc = await db
-    .collection<WithId<ITodo>>("todo")
-    .findOne({ _id: new ObjectId(todoid) });
+  try {
+    if (!todoid || typeof todoid !== "string" || !ObjectId.isValid(todoid)) {
+      throw new Error(`Invalid ObjectId Type ${todoid}`);
+    }
 
-  if (!todoDoc) {
-    return { message: "조회 결과 해당 투두가 없습니다." };
+    const db = (await connectDB).db("next-todo-chart-cluster");
+    const todoDoc = await db
+      .collection<WithId<ITodo>>("todo")
+      .findOne({ _id: new ObjectId(todoid) });
+
+    if (!todoDoc) {
+      throw new Error("Todo not found");
+    }
+    if (todoDoc.userid !== userid) {
+      return { message: "투두를 작성한 사용자만 투두를 삭제할 수 있습니다." };
+    }
+
+    await db.collection("todo").deleteOne({ _id: new ObjectId(todoid) });
+    await db.collection("todos").updateOne(
+      {
+        author: userid,
+      },
+      {
+        $pull: { content: new ObjectId(todoid) },
+      },
+    );
+
+    revalidateTag(`todo-${todoid}`);
+    revalidateTag("todos");
+
+    return { message: "" };
+  } catch (error) {
+    if (error instanceof Error) {
+      return {
+        message: `투두 추가 과정 중 에러가 발생했습니다. ${error.message}`,
+      };
+    }
+    console.error(error);
+    return { message: "" };
   }
-  if (todoDoc.userid !== userid) {
-    return { message: "투두를 작성한 사용자만 투두를 삭제할 수 있습니다." };
-  }
-
-  await db.collection("todo").deleteOne({ _id: new ObjectId(todoid) });
-  await db.collection("todos").updateOne(
-    {
-      author: userid,
-    },
-    {
-      $pull: { content: new ObjectId(todoid) },
-    },
-  );
-
-  revalidateTag(`todo-${todoid}`);
-  revalidateTag("todos");
-
-  return { message: "" };
 };
